Fix spark-sql-perf for broader usage.
spark-sql-perf requires that your Datasets be stored in the default Hive
Database. My change adds flags to use an Spark SQL datasource and/or an
HCFS directory containing the data and creates temporary Views in the
default database

It also makes some small changes to the TPCDS Benchmark and Query
discovery reflection to allow it to run the TPCDS 2.4 queries out of the
box.

The BenchmarkRunner class was slightly negelected as Databricks seems to
prefer running in a notebook as opposed to a precompiled binary.